Add department and user functionalities, modify database relationships
Added multiple new properties to the User domain model, such as CompanyId, DepartmentId, LocationId, RoleId, and their corresponding navigation properties to establish a relationship. Also added a method to update department status in the DepartmentController. Updated database relationships, eliminated their direct relationship with the Customer entity, and created a direct relationship with the User entity. This change was made to ensure data integrity and encapsulation since a user now has a department, a role and is part of a company and a location.