Refactor database relations and constraints
This commit removes the direct relationship between Customer and Company, Location, Department, and Role entities, and instead adds customerId as a foreign key to Department and Location entities. It further sets a one-to-one relationship between User and Company and many-to-one relationships between User and the other three entities. The reason behind this restructuring is to align more with the business logic where a company has many users who can belong to many departments and locations. This makes data handling more streamlined and logical.