Add initial database migrations and update Location update logic
Added a new initial database migration which creates multiple tables for the RDF.Arcana.API project. These tables are structured according to the defined models and include companies, customers, departments, items, locations, meat types, product sub categories and categories, roles, UOMs and users. This establishes the initial database schema for the project.

Also, updated the location update logic to throw an exception if a location is not found. This change was made to improve error handling.

This commit