"Renamed Role to UserRoles in User domain and updated related files
User domain's RoleId attribute and Role navigation property have been renamed to UserRoleId and UserRoles respectively, in line with a modification to the Role domain. The corresponsing mapper in the Users feature, where User domain instances are converted to DTOs, has been adjusted accordingly. Moreover, a new feature for updating role status was added and the User domain now includes a field for permissions associated with a role. Certain attributes and methods in other areas have also been modified in conjunction with these changes. This renaming of Role to UserRoles better reflects how they are applied to users, ensuring improved clarity and maintainability."