Remove redundant functionality in the API
Deleted set of backend API features and related domain classes (AddMainMenu, AddNewModule, AddNewRole and others) which provided functionality to add, update, and retrieve role, module and main menu in the system. These features were removed due to a shift in project requirements and design, rendering these specific functions unnecessary to the overall workflow of the application.